L8SportsManagementSystem Fix multiple bugs in login, database connection, entire structure of the code is changed to work properly.
