+++ L8SportsManagementSystem/composer.json
@@ -13,5 +13,8 @@
     ],
     "require-dev": {
         "phpunit/phpunit": "^9"
+    },
+    "require": {
+        "vlucas/phpdotenv": "^5.6"
     }
 }
