--- L8SportsManagementSystem/config/Database.php
@@ -1,36 +1,0 @@
-<?php
-
-class Database
-{
-    private $host;
-    private $db_name;
-    private $username;
-    private $password;
-    public $conn;
-
-    public function __construct()
-    {
-        $this->host = getenv('DB_HOST');
-        $this->db_name = getenv('DB_NAME');
-        $this->username = getenv('DB_USER');
-        $this->password = getenv('DB_PASS');
-    }
-
-    public function getConnection()
-    {
-        $this->conn = null;
-
-        try {
-            $this->conn = new PDO(
-                'mysql:host=' . $this->host . ';dbname=' . $this->db_name,
-                $this->username,
-                $this->password
-            );
-            $this->conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-        } catch (PDOException $e) {
-            echo 'Connection Error: ' . $e->getMessage();
-        }
-
-        return $this->conn;
-    }
-}
