--- L8SportsManagementSystem/public/css/styles.css
@@ -1,26 +1,0 @@
-/* assets/style.css */
-#sidebar {
-    min-width: 200px;
-    max-width: 200px;
-    height: 100vh;
-    position: fixed;
-}
-
-#main-content {
-    margin-left: 220px; /* Sidebar width + 20px padding */
-}
-
-@media (max-width: 768px) {
-    #sidebar {
-        min-width: 100px;
-        max-width: 100px;
-    }
-
-    #main-content {
-        margin-left: 120px;
-    }
-}
-
-.nav-link {
-    font-size: 1.1rem;
-}
