+++ L8SportsManagementSystem/src/Database.php
@@ -1,0 +1,29 @@
+<?php
+
+class Database
+{
+    private $host;
+    private $db_name;
+    private $username;
+    private $password;
+    public $conn;
+
+    public function __construct()
+    {
+        $this->host = $_ENV['DB_HOST'];
+        $this->db_name = $_ENV['DB_NAME'];
+        $this->username = $_ENV['DB_USER'];
+        $this->password = $_ENV['DB_PASS'];
+    }
+
+    public function getConnection()
+    {
+        $this->conn = new \mysqli($this->host, $this->username, $this->password, $this->db_name);
+
+        if ($this->conn->connect_error) {
+            die("Connection failed: " . $this->conn->connect_error);
+        }
+
+        return $this->conn;
+    }
+}
