+++ L8SportsManagementSystem/src/GameController.php
@@ -1,7 +1,12 @@
 <?php
 
-require_once '../../config/Database.php';
-require_once '../Models/Game.php';
+require_once '../vendor/autoload.php';
+
+$dotenv = Dotenv\Dotenv::createImmutable(__DIR__ . '/../'); // Adjust path as necessary
+$dotenv->load();
+
+require_once 'Database.php';
+require_once 'Game.php';
 
 $database = new Database();
 $conn = $database->getConnection();
