+++ L8SportsManagementSystem/src/TeamController.php
@@ -1,7 +1,12 @@
 <?php
 
-require_once '../../config/Database.php';
-require_once '../Models/Team.php';
+require_once '../vendor/autoload.php';
+
+$dotenv = Dotenv\Dotenv::createImmutable(__DIR__ . '/../'); // Adjust path as necessary
+$dotenv->load();
+
+require_once 'Database.php';
+require_once 'Team.php';
 
 $database = new Database();
 $conn = $database->getConnection();
