+++ L8SportsManagementSystem/src/css/styles.css
@@ -1,0 +1,43 @@
+#sidebar {
+    min-width: 200px;
+    max-width: 200px;
+    height: 100vh;
+    position: fixed;
+}
+
+.page {
+    position: relative;
+    left: 200px;
+    margin-left: 50px;
+    margin-top: 50px;
+    margin-right: 50px;
+}
+
+.background-container {
+    background-image: url('../images/sports-background.jpg');
+    background-size: cover;
+    background-position: center;
+    background-repeat: no-repeat;
+    background-attachment: fixed;
+}
+
+#main-content {
+    margin-left: 220px;
+}
+
+@media (max-width: 768px) {
+    #sidebar {
+        min-width: 100px;
+        max-width: 100px;
+    }
+
+    #main-content {
+        margin-left: 120px;
+    }
+}
+
+.nav-link {
+    font-size: 1.1rem;
+}
+
+
