+++ L8SportsManagementSystem/src/login.php
@@ -1,0 +1,48 @@
+<?php
+
+require_once '../vendor/autoload.php';
+
+$dotenv = Dotenv\Dotenv::createImmutable(__DIR__ . '/../'); // Adjust path as necessary
+$dotenv->load();
+
+require_once 'Database.php';
+require 'AuthController.php';
+
+if (session_status() == PHP_SESSION_NONE) {
+    session_start();
+}
+
+$authController = new AuthController;
+
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    $authController->login($_POST['username'], $_POST['password']);
+}
+?>
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Login - Sports Management</title>
+    <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
+    <link href="./css/styles.css" rel="stylesheet">
+</head>
+<body class="d-flex justify-content-center align-items-center vh-100 bg-light background-container">
+<div class="">
+    <div class="card p-4 shadow-lg" style="width: 350px;">
+        <h3 class="card-title text-center mb-3">Login</h3>
+        <form action="login.php" method="post">
+            <div class="form-group">
+                <label for="username">Username</label>
+                <input type="text" name="username" class="form-control" required>
+            </div>
+            <div class="form-group">
+                <label for="password">Password</label>
+                <input type="password" name="password" class="form-control" required>
+            </div>
+            <button type="submit" class="btn btn-primary btn-block">Login</button>
+        </form>
+    </div>
+</div>
+</body>
+</html>
