+++ L8SportsManagementSystem/src/logout.php
@@ -1,0 +1,8 @@
+<?php
+
+require_once '../vendor/autoload.php';
+require 'AuthController.php';
+
+$authController = new AuthController;
+
+$authController->logout();
